+++ frontend/src/app/app.module.ts
@@ -1,8 +1,8 @@
-import {NgModule} from '@angular/core';
+import {NgModule, OnInit} from '@angular/core';
 import {BrowserModule} from '@angular/platform-browser';
 
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
-import {EntryComponent} from './components/entry/entry.component';
+import {EntryComponent} from './modules/entry/entry.component';
 import {FormsModule, ReactiveFormsModule} from "@angular/forms";
 import {HttpClientModule} from "@angular/common/http";
 
@@ -13,7 +13,11 @@ import {DialogContentSignupDialog, SignupComponent} from './components/authentic
 import {NgbModule} from '@ng-bootstrap/ng-bootstrap';
 import {ToastrModule} from "ngx-toastr";
 import {DialogContentSigninDialog, SigninComponent} from './components/authentication/signin/signin.component';
-import {DialogContentProfileEditDialog, ProfileComponent} from './components/profile/profile.component';
+import {
+  DialogContentProfileEditDialog,
+  DialogContentProfileEnterGameDialog,
+  ProfileComponent
+} from './modules/profile/profile.component';
 import {LoadingIndicatorComponent} from './components/utility/loading-indicator/loading-indicator.component';
 import {MatChipsModule} from "@angular/material/chips";
 import {MatCardModule} from "@angular/material/card";
@@ -26,7 +30,19 @@ import {MatProgressSpinnerModule} from "@angular/material/progress-spinner";
 import {MatAutocompleteModule} from "@angular/material/autocomplete";
 import {MatOptionModule} from "@angular/material/core";
 import {MatIconModule} from "@angular/material/icon";
-import { ExperimentalComponent } from './components/experimental/experimental.component';
+import { ExperimentalComponent } from './modules/experimental/experimental.component';
+import {OverlayContainer} from "@angular/cdk/overlay";
+import {MatMenuModule} from "@angular/material/menu";
+import { LobbyComponent } from './modules/lobby/lobby.component';
+import { WerewolfLobbyComponent } from './modules/lobby/werewolf-lobby/werewolf-lobby.component';
+import {DragDropModule} from "@angular/cdk/drag-drop";
+import {MatListModule} from "@angular/material/list";
+import {MatProgressBarModule} from "@angular/material/progress-bar";
+import {MatTabsModule} from "@angular/material/tabs";
+import { RoleSelectionComponent } from './components/lobby/role-selection/role-selection.component';
+import { PlayerListComponent } from './components/lobby/player-list/player-list.component';
+
+const THEME_KEY = 'theme-style';
 
 @NgModule({
   declarations: [
@@ -38,8 +54,13 @@ import { ExperimentalComponent } from './components/experimental/experimental.co
     DialogContentSigninDialog,
     ProfileComponent,
     DialogContentProfileEditDialog,
+    DialogContentProfileEnterGameDialog,
     LoadingIndicatorComponent,
-    ExperimentalComponent
+    ExperimentalComponent,
+    LobbyComponent,
+    WerewolfLobbyComponent,
+    RoleSelectionComponent,
+    PlayerListComponent
   ],
   imports: [
     BrowserModule,
@@ -60,11 +81,32 @@ import { ExperimentalComponent } from './components/experimental/experimental.co
     MatProgressSpinnerModule,
     MatAutocompleteModule,
     MatOptionModule,
-    MatIconModule
+    MatIconModule,
+    MatMenuModule,
+    DragDropModule,
+    MatListModule,
+    MatProgressBarModule,
+    MatTabsModule
   ],
   entryComponents: [DialogContentSignupDialog, DialogContentSigninDialog, DialogContentProfileEditDialog],
   providers: [authInterceptorProviders],
   bootstrap: [AppComponent]
 })
 export class AppModule {
+
+  constructor(private overlayContainer: OverlayContainer) {
+  }
+
+  changeTheme(theme: 'light-theme' | 'dark-theme'): void {
+    const overlayContainerClasses = this.overlayContainer.getContainerElement().classList;
+
+    const themeClassesToRemove = Array.from(overlayContainerClasses)
+      .filter((item: string) => item.includes('-theme'));
+    if (themeClassesToRemove.length) {
+      overlayContainerClasses.remove(...themeClassesToRemove);
+    }
+
+    overlayContainerClasses.add(theme);
+  }
+
 }