+++ frontend/src/app/components/utility/loading-indicator/loading-indicator.component.spec.ts
@@ -22,4 +22,5 @@ describe('LoadingIndicatorComponent', () => {
   it('should create', () => {
     expect(component).toBeTruthy();
   });
+
 });