+++ frontend/src/app/entities/voicepack.ts
@@ -1,0 +1,37 @@
+export abstract class VoicePack {
+
+}
+
+export class VoiceLine {
+
+  constructor() {
+  }
+
+  public codenameAlfa: VoiceLine[];
+  public codenameBravo: VoiceLine[];
+  public codenameCharlie: VoiceLine[];
+  public codenameDelta: VoiceLine[];
+  public codenameEcho: VoiceLine[];
+  public codenameFoxtrot: VoiceLine[];
+  public codenameGolf: VoiceLine[];
+  public codenameHotel: VoiceLine[];
+  public codenameIndia: VoiceLine[];
+  public codenameJuliett: VoiceLine[];
+  public codenameKilo: VoiceLine[];
+  public codenameLima: VoiceLine[];
+  public codenameMike: VoiceLine[];
+  public codenameNovember: VoiceLine[];
+  public codenameOscar: VoiceLine[];
+  public codenamePapa: VoiceLine[];
+  public codenameQuebec: VoiceLine[];
+  public codenameRomeo: VoiceLine[];
+  public codenameSierra: VoiceLine[];
+  public codenameTango: VoiceLine[];
+  public codenameUniform: VoiceLine[];
+  public codenameVictor: VoiceLine[];
+  public codenameWhiskey: VoiceLine[];
+  public codenameXray: VoiceLine[];
+  public codenameYankee: VoiceLine[];
+  public codenameZulu: VoiceLine[];
+
+}