+++ frontend/src/app/enums/playerstate.ts
@@ -1,0 +1,4 @@
+export enum PlayerState {
+  ALIVE = "ALIVE",
+  DECEASED = "DECEASED"
+}